# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Centre Commercial API
  description: |
    API de gestion de centre commercial (MEAN stack).

    ## Rôles
    - **ADMIN** : gestion complète (boutiques, catégories, utilisateurs, stats)
    - **SHOP_MANAGER** : gestion de sa propre boutique (produits, promotions)
    - **CUSTOMER** : navigation, favoris, avis

    ## Authentification
    Stratégie double token JWT :
    - **Access token** : 15 min — à envoyer dans `Authorization: Bearer <token>`
    - **Refresh token** : 7 jours — à envoyer dans le body pour rafraîchir

    Les tokens ne sont délivrés qu'après vérification de l'email.
  version: '1.0.0'
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
security:
  - BearerAuth: []
...
paths:
...
  /auth/login-mfa:
    post:
      tags: [Auth]
      security: []
      operationId: loginMfa
      summary: Connexion MFA — étape 1
      description: Valide les identifiants et envoie un code à 6 chiffres par email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Code MFA envoyé par email
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user_id:
                            type: string
                            description: ID à fournir à /auth/verify-mfa-code
  /auth/verify-mfa-code:
    post:
      tags: [Auth]
      security: []
      operationId: verifyMfaCode
      summary: Connexion MFA — étape 2
      description: Vérifie le code reçu par email et délivre les tokens JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, code]
              properties:
                user_id:
                  type: string
                code:
                  type: string
                  example: '123456'
      responses:
        '200':
          description: Authentification MFA réussie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-email/{token}:
    post:
      tags: [Auth]
      security: []
      operationId: verifyEmail
      summary: Vérification de l'adresse email
      description: Valide le token reçu par email et délivre les JWT.
      parameters:
        - $ref: '#/components/parameters/TokenParam'
      responses:
        '200':
          description: Email vérifié, tokens délivrés
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh-token:
    post:
      tags: [Auth]
      security: []
      operationId: refreshToken
      summary: Rafraîchir l'access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Nouveau access token généré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags: [Auth]
      security: []
      operationId: forgotPassword
      summary: Demande de réinitialisation de mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email de réinitialisation envoyé (même réponse si l'email n'existe pas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
  /auth/reset-password/{token}:
    post:
      tags: [Auth]
      security: []
      operationId: resetPassword
      summary: Réinitialisation du mot de passe
      parameters:
        - $ref: '#/components/parameters/TokenParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, confirmPassword]
              properties:
                password:
                  type: string
                  format: password
                confirmPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Mot de passe réinitialisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /auth/profile:
    put:
      tags: [Auth]
      operationId: updateProfile
      summary: Modifier son profil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TokenParam:
      name: token
      in: path
      required: true
      schema:
        type: string
      description: Token de vérification (brut, non haché)
  responses:
    Unauthorized:
      description: Non authentifié — token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: Données invalides (validation échouée)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
    UserRole:
      type: string
      enum: [CUSTOMER, SHOP_MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [VALID, SUSPENDED, WAITING]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: Valide 15 minutes
        refreshToken:
          type: string
          description: Valide 7 jours
        user:
          $ref: '#/components/schemas/UserProfile'
    UpdateProfileInput:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        address:
          type: string

    # --- Users ---
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        address:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
...